separated test-case which is not yet PdoQueryReflector supported
